pyproject.toml
[project]
name = "py-leap"
version = "0.1a41-dev"
description = "Antelope protocol framework"
authors = [{ name = "Guillermo Rodriguez", email = "guillermo@telos.net" }]
requires-python = ">=3.11"
readme = "README.md"
license = "AGPL-3.0-or-later"
dependencies = [
    "requests<2.32.0",
    "httpx>=0.28.1,<0.29",
    "msgspec>=0.19.0",
    "pyo3-antelope-rs",
    "trio-websocket>=0.12.2",
    "websockets>=15.0.1",
    "tractor",
    "psutil>=7.0.0",
    # linux x86_64
    "pyo3-antelope-rs @ https://github.com/openrepublic/pyo3-antelope-rs/releases/download/v0.8.1/pyo3_antelope_rs-0.8.1-cp39-abi3-manylinux_2_34_x86_64.whl ; sys_platform == 'linux' and platform_machine == 'x86_64'",
    # windows amd64
    "pyo3-antelope-rs @ https://github.com/openrepublic/pyo3-antelope-rs/releases/download/v0.8.1/pyo3_antelope_rs-0.8.1-cp39-abi3-win_amd64.whl ; sys_platform == 'win32' and platform_machine == 'AMD64'",
    # macos x86_64
    "pyo3-antelope-rs @ https://github.com/openrepublic/pyo3-antelope-rs/releases/download/v0.8.1/pyo3_antelope_rs-0.8.1-cp39-abi3-macosx_10_12_x86_64.whl ; sys_platform == 'darwin' and platform_machine == 'x86_64'",
    # macos arm64
    "pyo3-antelope-rs @ https://github.com/openrepublic/pyo3-antelope-rs/releases/download/v0.8.1/pyo3_antelope_rs-0.8.1-cp39-abi3-macosx_11_0_arm64.whl ; sys_platform == 'darwin' and platform_machine == 'arm64'",
]

[dependency-groups]
docs = [
    "sphinx==7.1.2",
    "sphinx-rtd-theme==1.3.0",
]
dev = [
    "pytest>=8.3.4,<9",
    "docker>=6.1.3,<7",
    "greenback>=1.2.1",
    "pdbp>=1.6.1",
]
snaps = [
    "bs4>=0.0.2,<0.0.3",
    "tdqm>=0.0.1,<0.0.2",
    "zstandard>=0.21.0,<0.22",
]

[tool.uv]
default-groups = []
environments = [
    "sys_platform == 'linux' and platform_machine == 'x86_64'",
    "sys_platform == 'win32'  and platform_machine == 'AMD64'",
    "sys_platform == 'darwin' and platform_machine == 'x86_64'",
    "sys_platform == 'darwin' and platform_machine == 'arm64'",
]

[tool.uv.sources.tractor]
git = "https://pikers.dev/goodboy/tractor.git"
rev = "053078ce8fd52284d767b200e77bb2bd23b5c812"
# path = "../tractor"
# editable = true

# [tool.uv.sources.pyo3-antelope-rs]
# path = "../pyo3-antelope-rs"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.sdist]
include = ["src/leap"]

[tool.hatch.build.targets.wheel]
include = ["src/leap"]

[tool.hatch.build.targets.wheel.sources]
"src/leap" = "leap"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.entry-points."pytest11"]
leap = "leap.fixtures"

[tool.ruff]
exclude = [
    ".git",
    ".ruff_cache",
    ".venv",
]

line-length = 80
indent-width = 4
target-version = "py311"

[tool.pytest.ini_options]
addopts = [
  '--show-capture=no',
]
markers = [
    'bootstrap: test requires full chain bootstrap',
    'contracts: specify contract you need deployed',
    'randomize: randomize eosio key and ports',
]
